Update stemcell docs to include jammy
Trusty has been out of support for a while, so we don't need to mention
it in index.html.md.erb, and we can remove all the external pages (that
didn't to be linked to) that described Trusty version lines.

Many of the Xenial version lines in index.html.md.erb are not supported
anymore, removing those.

Xenial is no longer the new kid on the block, and
xenial-tiles.html.md.erb should be updated to reflect that. As well as
include documentation links that work. Since some of the release notes
for versions no longer exist, we should switch to a current
documentation link.

Refactor the get-stemcells script so that Xenial (multiple version
lines) and Jammy (only one version line right now) can share the same
code. Also the anchor links need to reflect the Ubuntu version so that
they don't clash in the future.

Regenerated the stemcells.html.md.erb file with the script locally.

Signed-off-by: Aram Price <pricear@vmware.com>